Increase precision of wait read/write select to accept floats

+++ src/libckpool.c
@@ -903,7 +903,7 @@ int wait_close(int sockd, int timeout)
 }
 
 /* Emulate a select read wait for high fds that select doesn't support */
-int wait_read_select(int sockd, int timeout)
+int wait_read_select(int sockd, float timeout)
 {
 	struct pollfd sfd;
 	int ret = -1;
@@ -985,7 +985,7 @@ out:
 }
 
 /* Emulate a select write wait for high fds that select doesn't support */
-int wait_write_select(int sockd, int timeout)
+int wait_write_select(int sockd, float timeout)
 {
 	struct pollfd sfd;
 	int ret = -1;
